requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.4
anyio==4.2.0
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.1
av==9.1.1
azure-cognitiveservices-speech==1.34.0
beautifulsoup4==4.11.1
Brotli==1.1.0
bs4==0.0.1
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
click-option-group==0.5.6
click-params==0.4.1
colorama==0.4.4
comtypes==1.2.1
cryptography==41.0.7
cssselect==1.2.0
cssselect2==0.7.0
decorator==4.4.2
defusedxml==0.7.1
discord==2.1.0
discord.py==2.1.0
distlib==0.3.6
elastic-transport==8.4.0
elasticsearch==5.5.3
exceptiongroup==1.2.0
fake-useragent==0.1.11
ffmpy==0.3.1
filelock==3.8.0
frozenlist==1.3.3
gobject==0.1.0
greenlet==3.0.1
gst==0.1.0
gTTS==2.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==5.0.0
Jinja2==3.1.2
joblib==1.3.2
keyboard==0.13.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mashumaro==3.11
mdurl==0.1.2
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mutagen==1.47.0
networkx==3.2.1
numba==0.57.1
numpy==1.24.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
outcome==1.2.0
packaging==23.2
parse==1.19.0
pastehunter==1.4.2
pefile==2023.2.7
phonenumbers==8.12.57
Pillow==10.1.0
pipenv==2022.10.25
platformdirs==2.5.2
playwright==1.40.0
pooch==1.8.0
proglog==0.1.10
protobuf==4.25.1
protobuf3-to-dict==0.1.5
psutil==5.8.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodomex==3.19.0
pyee==11.0.1
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pymusiclooper==3.2.2
pynput==1.7.6
pyOpenSSL==23.3.0
pyoxidizer==0.24.0
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
pyppeteer==1.0.2
pyquery==1.4.3
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytaglib==2.1.0
python-decouple==3.6
python-dotenv==1.0.0
python-vlc==3.0.20123
pyttsx3==2.90
pytweening==1.0.7
pywin32==306
pywin32-ctypes==0.2.2
qrcode==7.4.2
requests==2.27.1
requests-html==0.10.0
rich==13.7.0
rich-click==1.7.2
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.7
splunk-sdk==1.7.2
stem==1.8.1
svgwrite==1.4.3
sympy==1.12
threadpoolctl==3.2.0
TikTokApi==6.2.0
TikTokLive==5.0.8
tinycss2==1.2.1
tkvideoplayer==2.5
tls-client==0.2.2
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.8.0
undetected-chromedriver==3.5.4
urllib3==1.26.18
urllib3-secure-extra==0.1.0
validators==0.20.0
vboxapi==1.0
virtualenv==20.16.6
virtualenv-clone==0.5.7
w3lib==2.0.1
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==10.4
wget==3.2
wsproto==1.2.0
yara-python==4.2.3
yarl==1.8.2
yt-dlp==2023.11.16
zipp==3.10.0
